+++ src/components/header/MobileSearchOverlay.tsx
@@ -1,0 +1,88 @@
+'use client';
+
+import { useEffect, useRef, useState } from 'react';
+import { Search, X } from 'lucide-react';
+import { useTranslations } from 'next-intl';
+import { useRouter } from '@/i18n/navigation';
+import { cn } from '@/components/ui/utils';
+
+interface MobileSearchOverlayProps {
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+export function MobileSearchOverlay({ isOpen, onClose }: MobileSearchOverlayProps) {
+  const router = useRouter();
+  const t = useTranslations('header');
+  const [query, setQuery] = useState('');
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  useEffect(() => {
+    if (isOpen) {
+      setTimeout(() => inputRef.current?.focus(), 50);
+    } else {
+      setQuery('');
+    }
+  }, [isOpen]);
+
+  useEffect(() => {
+    const handleKey = (e: KeyboardEvent) => {
+      if (e.key === 'Escape') onClose();
+    };
+    document.addEventListener('keydown', handleKey);
+    return () => document.removeEventListener('keydown', handleKey);
+  }, [onClose]);
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    const trimmed = query.trim();
+    if (!trimmed) return;
+    router.push('/search?q=' + encodeURIComponent(trimmed));
+    onClose();
+  };
+
+  return (
+    <>
+      {/* 배경 딤 */}
+      <div
+        className={cn(
+          'md:hidden fixed inset-0 z-[45] bg-black/20 transition-opacity duration-300',
+          isOpen ? 'opacity-100 pointer-events-auto' : 'opacity-0 pointer-events-none',
+        )}
+        onClick={onClose}
+        aria-hidden="true"
+      />
+      {/* 검색 패널 — 헤더 바로 아래 페이드인 */}
+      <div
+        className={cn(
+          'md:hidden fixed left-0 right-0 z-[46] bg-background/95 backdrop-blur-md shadow-md transition-[opacity,transform] duration-200 ease-in-out',
+          isOpen ? 'opacity-100 pointer-events-auto translate-y-0' : 'opacity-0 pointer-events-none -translate-y-2',
+        )}
+        style={{ top: '48px' }}
+        role="search"
+        aria-label={t('searchLabel')}
+      >
+        <form onSubmit={handleSubmit} className="flex items-center gap-3 px-4 py-3 border-b border-divider-soft">
+          <Search className="h-5 w-5 text-muted-foreground shrink-0" />
+          <input
+            ref={inputRef}
+            type="search"
+            value={query}
+            onChange={(e) => setQuery(e.target.value)}
+            placeholder={t('searchShort')}
+            aria-label={t('searchLabel')}
+            className="flex-1 bg-transparent text-base text-foreground placeholder:text-muted-foreground focus:outline-none"
+          />
+          <button
+            type="button"
+            onClick={onClose}
+            aria-label={t('searchClose')}
+            className="p-1 text-muted-foreground hover:text-foreground transition-colors"
+          >
+            <X className="h-5 w-5" />
+          </button>
+        </form>
+      </div>
+    </>
+  );
+}
